Changes to permit application of update() to optmatch.dlists's
"call" attribute added to values of mdist() and caliper().  Only way I found to do this required modification both of the generic and of the specific methods.
This in turn required importing getCall generic from stats package.  It's possible that I didn't really need to add a dependency on stats (the change to DESCRIPTION); haven't checked this yet.
Tests had to be adjusted here and there, just so they'd run.  At the moment update() is only tested on values of mdist.glm, not other methods.