Moving match_on docs inline
There is a bug with Roxygen that it doesn't include the appropriate
alias declarations for R CMD check. This bug might be fixed in the S4
branch of Roxygen2, but that branch fails for other reasons when making
the docs, so we need to put up with the warnings for now.