Documentation, build process improvements
Added a big header to the Makefile to provide more details on the major tasks
used during development.

All dependencies from CRAN are now installed in .local, whether they exist on
the system or not. This is useful for getting up and running on a fresh box
with all the development dependencies that are not explicitly listed in the
DESCRIPTION file (e.g. testthat for running tests, RItools which is optional).