Using more general include path for R.h
Distributions are free to put the headers in non-standard locations as
long as the R_INCLUDE_DIR variable is properly set to point to the
location. Normally, this is identical to R_HOME/include, but in the case
of ArchLinux, the location is /usr/include/R. The makefile.linux should
now be more flexible.