Removing "as.matrix" from caliper examples.
The calls to match_on and caliper were wrapped through as.matrix, "for convenient presentation only".  But on my machine I couldn't see any difference between what that produced and what I got without the as.matrix. So I'm assuming that this is a relic of some earlier version of the print methods for ISMs, and taking out the as.matrix'es. Do revert if I am mistaken.