hasty solution to malformed error message
This fixes an issue due to long x'es being split into
multiple entries in a character vector. As an example,
an x of this form was fed to fullmatch
    psDist2+caliper(psDist2,4)
    +caliper(absdist,2)
    +caliper(mahalDist,40)
(This input was mistaken b/c `psDist2` did not exist.)
The error message that was generated read
    Error in value[[3L]](cond): Can't find psDist2 + caliper(psDist2, 4) + caliper(absdist, 2) + caliper(mahalDist, . If it lives within the data frame provided
    as the data argument, try
    meddat$psDist2 + caliper(psDist2, 4) + caliper(absdist, 2) + caliper(mahalDist,
    Can't find   40).
    If it lives within the data frame
    provided as the data argument, try meddat$    40)

which doesn't make sense due to x_str being a character vector of length 2,
with the final "`40)`" having wrapped to its second entry.