MCFSolutions c() method skips validity check on combined object
--in order to save time on larger problems.  However,
it will perform separate validity checks on combined versions of
the constituent objects nodes (of class NodeInfo), arcs (ArcInfo),
etc. What it doesn't do is make sure these things are consistent
with one another.  That validity check will be much faster to
perform separately on the subproblems, prior to combining.

Faster to check
 validity on each of the smaller problems, which in most