Revert "Pass disthash+hint down call stack from fullmatch.matrix()"
(The idea of passing down the combination of the disthash and the hint
was to use them to inform decisions at the subproblem level as to whether
a new attempt at solving the problem needs to be made, vs whether the
last one was good enough.   But along the way I realized that it's likely
to be expensive to inspect the hint info to decide if the problem that
had been solved was similar in relevant respects. It seems a more promising
approach to design and implement these checks inside of update.fullmatch()/
update.pairmatch() etc, probably at the problem rather than the subproblem
level.)

This reverts commit 4dbf103860c122764b65d830f6305064f0d4c9e5.