Remove "How closely did I match?" subsec
We were recommending the use of `summary.optmatch()`
to assemble matched distance info, but as of this writing
it doesn't do that, and the subsection is internally incoherent
(promises matched distance information that the command we present
does not return). We may reenable that when addressing #173,
at which time we should consider reverting this commit.

If on the other hand we resolve #173 without restoring
sum-of-matched-distances information to the output of
`summary.optmatch()`, then it becomes more important to add
vignette material explaining how to use `matched.distances()`
to extract this info.  Cf. #185.