Update test code for R 4.0
We needed `groups` column to be a factor, and `data.frame()`
does this implicitly.

Incidentally, my message with commit ee72f52 said this branch caused
R version 4.0 to segfault. That was a false alarm; I had neglected to
`make clean` my working copy after upgrading R.  With this commit
all's good on R 4.0, so far as I know.