Address scores() mishandling of svyglm's
svyglm's leave a '(weights)' column in their model.frame's even
if not invoked with a weights= argument; they typically receive
weights through their `design=` argument.

To me it seemed probably safest in general to
look to `object`'s original call to decide whether or
not an explicit weights argument is appropriate. This this commit.