Address oversights
Logic branches for infeasible subproblems being identified as such
before passing to the solver needed output formats updated (in
`fmatch()` and in `solve_reg_fm_prob()`).  This was a quickie
version of that update that doesn't take the trouble to generate
an MCFSolution representing the situation, passing a NULL instead.